scrapper/binance_producer/requirements.txt
kafka-python-ng==2.2.3
websocket-client==1.6.2
requests==2.32.3